Update:taskテーブルのresourceカラムをtime_startとtime_endに変更しました

+++ db/migrate/20220504024525_create_schedules.rb
@@ -2,7 +2,7 @@ class CreateSchedules < ActiveRecord::Migration[7.0]
   def change
     create_table :schedules do |t|
       t.date :title, null: false
-      t.float :analysis
+      t.integer :analysis
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
